requirements.txt
# core libraries
torch==1.10.0 # (tensor ops and neural network modules)
einops==0.4.0 # (einstein operators)
dgl-cu113==0.7.2 # (graph neural network library)
networkx[all]==2.6.3 # (general purpose graph library)
tokenizers==0.10.3 # (efficient tokenizer implementations)
numpy==1.22.1 # (general math)

# others
ftfy==6.0.3 # (clean unicode errors from text)
lmdb==1.3.0 # (fast database to store large datasets)
lz4==3.1.10 # (compression)
spacy[cu113]==3.2.1 # (data preprocessing)
omegaconf==2.1.1 # (configuration of experiments)
tensorboard==2.7.0 # (experiment tracking)
python-Levenshtein==0.12.2 # (fast editdistance and editops calcuation in C)

# baselines
hunspell==0.5.5 # (hunspell spell checker)
jamspell==0.0.12 # (jamspell spell checker)
neuspell==1.0.0 # (neuspell spell checker suite)
symspellpy==6.7.6 # (python implementation of symspell algorithm)

# visualisation and demo
tqdm>=4.49.0 # (progress bars)
matplotlib==3.4.3 # (plotting)
streamlit==1.4.0 # (demo)
tabulate==0.8.9 # (markdown tables)
GPUtil==1.4.0 # (GPU stats)

# code style checks
pycodestyle==2.8.0 # (code style and formatting)
mypy==0.931 # (static type checking)

# testing
pytest==6.2.5 # (unit testing)
pytest-xdist==2.5.0 # (plugin to run pytest in parallel)
